refactor(api): contract cleanup + JWT-based authz for diagnosis APIs
Contract cleanup (TypeSpec):
- Uuid scalar with format pattern replaces bare string ids
- ListResponse<T> / ItemsResponse<T> shared shapes replace per-model
  list responses
- error-union aliases (AuthedErrors, NotFoundErrors, ...) replace
  hand-listed unions on every route

Diagnosis API authorization:
- work-values / career-interest sessions take the user from the JWT
  instead of a user_id in the request body (WVStartSessionRequest
  removed from the contract); writes require the owner's JWT
- new AnyJWTAuth middleware lets either candidate or company JWTs read
  diagnosis results
- team-diagnose sessions for invited (not logged-in) members are
  authorized by the invite token

Generated code (openapi.yaml, server.gen.go, sdk.gen.ts, types.gen.ts)
regenerated; frontend callers updated to the new SDK shapes.

+++ backend/internal/adapter/http/presenter/messaging_presenter.go
@@ -20,7 +20,7 @@ func MessagingConversationsResponse(convs []*messaging.ConversationWithPreview, 
 	for _, c := range messagingConv.ToConversationResponses(convs) {
 		items = append(items, *c)
 	}
-	return &openapi.ModelsConversationListResponse{Items: items, Total: total}
+	return &openapi.ModelsConversationListResponse{Items: items, Total: int32(total)}
 }
 
 // MessagingMessageResponse builds the single-message API response.
@@ -34,7 +34,7 @@ func MessagingMessagesResponse(msgs []*messaging.Message, total int) any {
 	for i, m := range msgs {
 		items[i] = *toMessageResponse(m)
 	}
-	return &openapi.ModelsMessageListResponse{Items: items, Total: total}
+	return &openapi.ModelsMessageListResponse{Items: items, Total: int32(total)}
 }
 
 // MessagingUnreadCountResponse builds the unread-count API response.
